Added a new NumberValue interface and added the other number value interfaces

+++ src/main/java/com/stardevllc/beans/value/DoubleValue.java
@@ -1,26 +1,26 @@
 package com.stardevllc.beans.value;
 
-public interface ByteValue extends NumberValue<Byte> {
-    byte get();
+public interface DoubleValue extends NumberValue<Double> {
+    double get();
     
     @Override
     default byte byteValue() {
-        return get();
+        return (byte) get();
     }
     
     @Override
     default short shortValue() {
-        return get();
+        return (short) get();
     }
     
     @Override
     default int intValue() {
-        return get();
+        return (int) get();
     }
     
     @Override
     default long longValue() {
-        return get();
+        return (long) get();
     }
     
     @Override
@@ -30,6 +30,6 @@ public interface ByteValue extends NumberValue<Byte> {
     
     @Override
     default float floatValue() {
-        return get();
+        return (float) get();
     }
 }
